update revenue api

+++ Common/ViewModels/RevenueStatisticViewModel.cs
@@ -9,6 +9,6 @@ namespace Common.ViewModels
     public class RevenueStatisticViewModel
     {
         public DateTime Date{ get; set;}
-        public double Revenue { get; set; }
+        public decimal Revenue { get; set; }
     }
 }
